chore: rename hooks; cleanup
