+++ src/app/segments/Segment.tsx
@@ -16,7 +16,7 @@ import DeleteSegmentPanel from '~/components/segments/DeleteSegmentPanel';
 import SegmentForm from '~/components/segments/SegmentForm';
 import Slideover from '~/components/Slideover';
 import { getSegment } from '~/data/api';
-import useError from '~/data/hooks/errors';
+import { useError } from '~/data/hooks/useError';
 import {
   ComparisonType,
   ConstraintOperators,
