+++ src/components/ErrorNotification.tsx
@@ -1,6 +1,6 @@
 import { Transition } from '@headlessui/react';
 import { XCircleIcon, XMarkIcon } from '@heroicons/react/20/solid';
-import useError from '~/data/hooks/errors';
+import { useError } from '~/data/hooks/useError';
 
 export default function ErrorNotification() {
   const { error, clearError } = useError();
