+++ src/components/Pagination.tsx
@@ -2,7 +2,7 @@ import {
   ArrowLongLeftIcon,
   ArrowLongRightIcon
 } from '@heroicons/react/20/solid';
-import { usePagination } from '~/data/hooks/pagination';
+import { usePagination } from '~/data/hooks/usePagination';
 import { classNames } from '~/utils/helpers';
 
 type PageProps = {
