+++ src/components/segments/ConstraintForm.tsx
@@ -8,7 +8,7 @@ import Input from '~/components/forms/Input';
 import Select from '~/components/forms/Select';
 import MoreInfo from '~/components/MoreInfo';
 import { createConstraint, updateConstraint } from '~/data/api';
-import useError from '~/data/hooks/errors';
+import { useError } from '~/data/hooks/useError';
 import { requiredValidation } from '~/data/validations';
 import {
   ComparisonType,
