+++ src/components/segments/DeleteSegmentPanel.tsx
@@ -2,7 +2,7 @@ import { Dialog } from '@headlessui/react';
 import { ExclamationTriangleIcon } from '@heroicons/react/24/outline';
 import Button from '~/components/forms/Button';
 import { deleteSegment } from '~/data/api';
-import useError from '~/data/hooks/errors';
+import { useError } from '~/data/hooks/useError';
 
 type DeleteSegmentPanelProps = {
   setOpen: (open: boolean) => void;
