+++ src/data/hooks/useError.ts
@@ -1,7 +1,7 @@
 import { useContext } from 'react';
 import { ErrorContext } from '~/components/ErrorProvider';
 
-export default function useError() {
+export const useError = () => {
   const { error, setError, clearError } = useContext(ErrorContext);
   return { error, setError, clearError };
-}
+};
