+++ vite.config.ts
@@ -11,14 +11,14 @@ export default defineConfig({
     }
   },
   build: {
-    manifest: true,
+    manifest: true
   },
   envPrefix: 'FLIPT_',
   server: {
     proxy: {
       '/api/v1': 'http://localhost:8080',
       '/auth/v1': 'http://localhost:8080',
-      '/meta': 'http://localhost:8080',
-    },
+      '/meta': 'http://localhost:8080'
+    }
   }
 });
